+++ CHANGELOG.md
@@ -1,12 +1,21 @@
 # Code Editor Package for Rider
-## [2.0.0] - 2019-11-25
+
+## [2.0.1] - 2020-03-05
+
+Speed improvements,
+ProjectTypeGuids for unity-generated project
+Improve UI for Project Generation settings
+Changes in csc.rsp would cause project-generation
+Remove NoWarn 0169 from generated csproj
+Support custom JetBrains Toolbox installation location
+
+## [1.2.1] - 2019-12-09
 
 - Load optimised EditorPlugin version compiled to net 461, with fallback to previous version.
 - On ExternalEditor settings page: reorder Generate all ... after Extensions handled
 - Better presentation for Rider of some version in ExternalEditors list
 - Initial support for Code Coverage with dotCover plugin in Rider
 - Added support for Player Project generation
-- Breaking API changes
 
 ## [1.1.4] - 2019-11-21
  - Fix warning - unreachable code